Created new matcher for abbrev_matcher instead of overriding fuzzy_matcher
